Добавляет в провайдер проверку на online

+++ src/api/provider.js
@@ -1,25 +1,39 @@
+const isOnline = () => {
+  return window.navigator.onLine;
+};
+
 export default class Provider {
   constructor(api) {
     this._api = api;
   }
 
   getFilms() {
-    return this._api.getFilms();
+    if (isOnline()) {
+      return this._api.getFilms();
+    }
   }
 
   getComments(filmId) {
-    return this._api.getComments(filmId);
+    if (isOnline()) {
+      return this._api.getComments(filmId);
+    }
   }
 
   updateFilm(id, data) {
-    return this._api.updateFilm(id, data);
+    if (isOnline()) {
+      return this._api.updateFilm(id, data);
+    }
   }
 
   createComment(filmId, comment) {
-    return this._api.createComment(filmId, comment);
+    if (isOnline()) {
+      return this._api.createComment(filmId, comment);
+    }
   }
 
   deleteComment(commentId) {
-    return this._api.deleteComment(commentId);
+    if (isOnline()) {
+      return this._api.deleteComment(commentId);
+    }
   }
 }
